moves GenericError and ParsingError(s) under Analyzer namespace
