+++ src/Analyzer/ParsingError.php
@@ -2,15 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Arkitect\Rules;
+namespace Arkitect\Analyzer;
 
 class ParsingError
 {
-    /** @var string */
-    private $relativeFilePath;
+    private string $relativeFilePath;
 
-    /** @var string */
-    private $error;
+    private string $error;
 
     public function __construct(string $relativeFilePath, string $error)
     {
