+++ tests/Unit/Analyzer/FileParserTest.php
@@ -8,9 +8,9 @@ use Arkitect\Analyzer\ClassDescriptions;
 use Arkitect\Analyzer\DocblockTypesResolver;
 use Arkitect\Analyzer\FileParser;
 use Arkitect\Analyzer\FileVisitor;
+use Arkitect\Analyzer\GenericError;
+use Arkitect\Analyzer\ParsingErrors;
 use Arkitect\CLI\TargetPhpVersion;
-use Arkitect\Rules\GenericError;
-use Arkitect\Rules\ParsingErrors;
 use PhpParser\NodeTraverser;
 use PhpParser\NodeVisitor\NameResolver;
 use PHPUnit\Framework\TestCase;
